+++ packages/react-ui/src/app/features/ai/assistant/assistant-ui-chat.tsx
